requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pandas
nest_asyncio
sqlalchemy
psycopg2-binary
pypdf
openpyxl

# LlamaIndex Core & Google
llama-index
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-postgres
google-generativeai
llama-parse